network_topology_strategy: allocate_tablets_for_new_table: do not rely on token ownership
Base initial tablets allocation for new table
on the dc/rack topology, with randomization,
rather then on the token ring, to remove the dependency
on token ownership.

Not only this method is more efficient that circling
the token ring and attemting to insert te endpoints
to the natural_endpoint_tracker until the replication
factor per dc is done, inserting the endpoints with
rack awareness also balances the load across racks
in each data center.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>