utils: chunked_vector: migrate: do not destroy each migrated element
The elements are migrated to a new chunk one at a time
using the move constructor.  This should move any resources
associated with the elements so destroying each element is
not expected to free any memory.

Deferring the destructor calls till after migration completes
better utilized the instruction cache as move is done in a tight loop
and when finished, the current chunk is destroyed, calling the
destruct for each destroyed element, again in a tight loop.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>