repair: repair_tablets: select primary dc per tablet
Tablet allocation does not guarantee fairness of
the first replica in the replicas set across dcs.
The lack of this fix cause the following dtest to fail:
repair_additional_test.py::TestRepairAdditional::test_repair_option_pr_multi_dc

Use the tablet_map calc_primary_replica_map to get the
map of primary replica per tablet id to see if this node
is the primary replica for each tablet or not.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>