repair: repair_tablets: use calc_primary_replica_map
Tablet allocation does not guarantee fairness of
the first replica in the replicas set across dcs.
The lack of this fix cause the following dtest to fail:
repair_additional_test.py::TestRepairAdditional::test_repair_option_pr_multi_dc

Use the tablet_map calc_primary_replica_map to get a
map of primary replicas per tablet id (possibly within the
local dc) to see if this node is the primary replica for each tablet or not.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>