tablet_map: get_primary_replica: pre-sort the replicas
The order of nodes within each data center in the
replica set is randomized. In general, we cannot
assume any order, and that might lead to imbalance
of the primary replica.

This change first sorts all replicas using
their host_id and then selects the primary replica
as the ordinal tablet id modulo the size of the replica
set.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>