effective_replication_map: refactor resolve_endpoints
Make the free template function `resolve_endpoints`
a static method in effective_replication_map, and introduce
a variant of it that uses the objects _tmptr to
call the static method.

This is done in preparation for change the
abstract_replication_strategy interface
to return both host_id based replica sets
and inet_address based sets.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>