init: set_this_node via database
Currently, we set this_node host_id directly
in the shared_token_metadata topology instance.

However, the system_keyspace keeps a reference to
an effective_replica_map which topology
is not updated.

This is fixed in this patch by using database::set_this_node
that propagates the change to all shards and updates all
effective_replication_map:s.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>