storage_service: refactor replica_to_all_cores out to replica::database
Now that the database keeps a reference to the
effective_replication_map_factory, it can update
all keyspaces and tables when the shared token_metadata
changes.

Refactor those parts out of storage_service::replica_to_all_cores
to replica::database::replica_to_all_cores and keep only
the session management and triggering of tablet
split candidates in storage_service.

Note that managing the tablets state is also
a good candidate to move to replica::database
but it's out of scope of this series.
It may be considered as a follow up.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>