hint_sender: use get_replicas rather get_natural_endpoints
As part of scylladb/scylladb#6403, we would
like to deprecate the use of `get_natural_endpoints`
and friends and use `get_replicas` instead.

do_send_one_mutation already validates the `end_point_key`
which is the destination host_id and makes sure it
can be mapped into an inet_address, so it would be
just fine if that host already left the cluster
(as in the case of get_replicas vs. get_natural_endpoints).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>