config: add enable_tablets_feature
Separate the configuration for enabling the tablets
feature from the enablement of tablets when creating
new keyspaces.

This way, if `enable_tablets_feature` is set to true
on all nodes, `enable_tablets` just controls whether
tablets are enabled or disabled by default for new keyspaces.

If `enable_tablets` is true, tablets can be opted-out
using `CREATE KEYSPACE WITH tablets = { 'enabled': false }`
as it is today.

If `enable_tablets` is false, tablets can be opted-in
using `CREATE KEYSPACE WITH tablets = { 'enabled': true }`

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>