sstables: scylla_metadata: add original sstable generation
Keep a copy of the sstable uuid generation in the
scylla_metadata component.

If the SSTable happens to have a numerical generation
just create a new time-uuid generation and log a warning.

Dump this new attribute in scylla sstable dump tool.

And add a unit test to verify that the written (and then
loaded) original sstable generation matches the sstable's
generation.

Fixes scylladb/scylladb#20459

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>