view: get_view_natural_endpoint: implement rack-aware pairing for tablets
If the keyspace uses tablets and the dc replication factor
is a multiple of the number of racks in the dc, all racks will have
the same number of replicas (replication_factor / number_of_racks)
and the tablet replicas would be confined to their rack and so it is
safe for base/view pairing to be rack aware.

Enabled with the tablets_basic_rack_aware_view_pairing cluster
feature.

Fixes scylladb/scylladb#17147

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>