utils: composite_abort_source: tunnel exception_ptr in subscription
abort is requested by one of the added abort sources,
intercept the requested exception and pass it on
to the composite_abort_source abort_source, so the abort
source can be distinguished.  For example, one of the
abort sources could be abort_on_expiry that generates
timed_out_error, and another abour_source could a generic
one, that goes off on shutdown, generating abort_requested_error.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>