table: flush: wait_until_stopped
When the table is stopped, all compaction groups
are stopped, and as part of that, they are flushing
their memtables.

table::flush can't enter the _pending_flushes_phaser
when the table is stopped since it is already closed,
so instead, just await wait_until_stopped() to wait
on any ongoing flushes that were induced by stop().

Fixes #22243

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>