transport: server: hold the query processor gate
Make sure that the query processor is kept alive
while queries are being processed.

Use the service permit to keep the gate holder
until the permit is destroyed.

Note that currently we use an empty permit
for bouncing the process function onto a different
shard.  In this case we hold that shard's query
processor's gate using an automatic variable,
since the processing on the remote shard is
limited to this call.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>