service_permit: hold phased_barrier operation
Make sure that the query processor and storage proxy
are kept alive while queries are being processed.

Use the service permit to keep the phased_barrier holder
until the permit is destroyed.

Note that currently we use an empty permit
for bouncing the process function onto a different
shard.  In this case we hold that shard's query
processor's operation using an automatic variable,
since the processing on the remote shard is
limited to this call.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>