db: snapshot: backup_task: limit concurrency
Otherwise, once all the background tasks are created
we have no way to reorder the queue.

Apply a limit on the number of bytes uploaded
concurrently.  This keeps the load on the s3
upload stack at bay, while having the benefit
of allowing to prioritize the upload queue -
to be done in a following patch.

Refs #23239

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>