locator: token_metadata: clear_gently in background if needed when destroyed
We have a lot of places in the code where
a token_metadata_ptr is kept in an automatic
variable and destroyed when it leaves the scope.
since it's a referenced counted lw_shared_ptr,
the token_metadata object is rarely destroyed in
those cases, but when it is, it doesn't go through
clear_gently, and in particular its tablet_metadata
is not cleared gently, leading to inefficient destruction
of potentially many foreign_ptr:s.

This patch adds a background call to clear_gently in
~token_metadata, if it hasn't been already called.
token_metadata::clear_gently was modified to move
the _impl member so it can be detected in the destructor.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>