table: try_flush_memtable_to_sstable: pass owned ranges in sstable_writer_config
To be used to generate the sstable's owned ranges.

Note that we change schema_applier::commit_tables_and_views
and schema_applier::commit_on_shard to be async.
This is safe since they are called internally by
schema_applier::commit() which is already async,
and it used replica::database::lock_tables_metadata to
serialize changes.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>