compaction_manager: needs_cleanup: use sstable token_ranges
If the sstable has token_ranges in scylla_metadata
use them to see if the sstable needs cleanup, or wheather
the sstable token ranges are all contained in the
sorted_owned_ranges.

This is much more accurate than just relying on the
"default" sstable token range which is using only the
first and last keys.

Fixes #12215

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>