test: object_store: test_restore_with_streaming_scopes: limit rack-aware restore to rf==#racks
With numeric replication factor, we don't guarantee that
the arbitrary subset of the racks picked initially would
be picked again at restore time.

To achieve that, the test needs to be extended to support
rack lists, and make sure the same lists are used when initially
creating the keyspace and when restoring it.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>